Separate Instance/Solution to files
