+++ vrplib/Solution.py
@@ -1,0 +1,6 @@
+from typing import List, TypedDict
+
+
+class Solution(TypedDict, total=False):
+    routes: List[List[int]]
+    cost: float
